requirements.txt
Django==4.0.2
django-active-link==0.1.8
django-ckeditor==6.2.0
django-crispy==1.8.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
whitenoise==6.0.0
